+++ blocks/branchcarousel/branchcarousel.js
@@ -1,4 +1,4 @@
-import { CFApiCall, groupAllKeys } from '../../scripts/scripts.js';
+import { CFApiCall, groupAllKeys } from '../../scripts/common.js';
 import { setLocationObj } from '../moredetailsaddress/moredetailsaddress.js';
 
 export default async function decorate(block) {
