+++ blocks/calculatorsection/calculatorsection.js
@@ -1,4 +1,4 @@
-import { calculatorFlatStrLogic, CFApiCall, fetchAPI } from '../../scripts/scripts.js';
+import { calculatorFlatStrLogic, CFApiCall, fetchAPI } from '../../scripts/common.js';
 import { homeLoanCalcFunc } from '../emiandeligiblitycalc/homeloancalculators.js';
 import { renderCalculatorData } from './renderdatafunc.js';
 import { homeloanCalHTML } from '../homeloancalculatorv2/templatehtmlv2.js';
