+++ blocks/code/code.js
@@ -1,5 +1,6 @@
 import { decorateIcons } from "../../scripts/aem.js";
-import { decodeHtmlSymbols, decoratePlaceholder } from "../../scripts/scripts.js";
+import { decoratePlaceholder } from "../../scripts/scripts.js";
+import { decodeHtmlSymbols } from "../../scripts/common.js";
 import { mobileHeaderAnalytics } from "./code-analytics.js";
 import { decorateTable } from "./table.js";
 export default async function decorate(block) {
