+++ blocks/financialreports/financialreports.js
@@ -1,5 +1,5 @@
 import { ctaClickInteraction } from '../../dl.js';
-import { fetchAPI, getProps, renderHelper } from '../../scripts/scripts.js';
+import { fetchAPI, getProps, renderHelper } from '../../scripts/common.js';
 
 export default async function decorate(block) {
   const props = getProps(block);
