+++ blocks/keyfeatures/keyfeatures.js
@@ -1,6 +1,7 @@
 import { ctaClick, ctaClickInteraction, keyFeaturesInteraction } from '../../dl.js';
 import { autoLinkLangPath, fetchPlaceholders } from '../../scripts/aem.js';
 import { decoratePlaceholder, renderHelper, targetObject } from '../../scripts/scripts.js';
+import { renderHelper } from '../../scripts/common.js';
 
 export default async function decorate(block) {
   block.innerHTML = await decoratePlaceholder(block);
