+++ blocks/sarfaesiretail/sarfaesiretail.js
@@ -1,4 +1,5 @@
-import { CFApiCall, fetchAPI, targetObject } from '../../scripts/scripts.js';
+import { targetObject } from '../../scripts/scripts.js';
+import { CFApiCall, fetchAPI } from '../../scripts/common.js';
 
 export default async function decorate(block) {
   const cfURL = block.textContent.trim();
